requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.3
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
mysqlclient==2.2.5
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.1
